+++ client/web/src/auth/welcome/StartSearching.tsx
@@ -3,13 +3,13 @@ import React, { useEffect, useState } from 'react'
 
 import { Link } from '@sourcegraph/shared/src/components/Link'
 import { ErrorLike, isErrorLike } from '@sourcegraph/shared/src/util/errors'
+import { useSteps } from '@sourcegraph/wildcard'
 
 import { AuthenticatedUser } from '../../auth'
 import { eventLogger } from '../../tracking/eventLogger'
 import { UserExternalServicesOrRepositoriesUpdateProps } from '../../util'
 import { LogoAscii } from '../LogoAscii'
 import { getPostSignUpEvent, RepoSelectionMode } from '../PostSignUpPage'
-import { useSteps } from '../Steps'
 import { Terminal, TerminalTitle, TerminalLine, TerminalDetails, TerminalProgress } from '../Terminal'
 import { useExternalServices } from '../useExternalServices'
 import { useRepoCloningStatus } from '../useRepoCloningStatus'
