+++ client/wildcard/src/components/index.ts
@@ -9,3 +9,4 @@ export { Tabs, Tab, TabList, TabPanel, TabPanels } from './Tabs'
 export { SourcegraphIcon } from './SourcegraphIcon'
 export { Badge, ProductStatusBadge } from './Badge'
 export { Menu, MenuButton, MenuDivider, MenuHeader, MenuItem, MenuLink, MenuItems, MenuPopover } from './Menu'
+export { Step, Steps, StepList, StepPanels, StepPanel, StepActions, useSteps } from './Steps'
